fleet_manager/requirements.txt
discord.py==2.3.2
redis==5.0.4
aiohttp==3.9.5
python-dotenv==1.0.1
